build: include plugin name in copyright header
