+++ build-config/fragments/common.js
@@ -6,17 +6,6 @@ const datefns = require('date-fns');
 
 const rootDir = path.resolve(__dirname, '..', '..');
 const pckg = require(path.join(rootDir, 'package.json'));
-const date = datefns.format(new Date(), 'YYYY-MM-DD');
-
-// add JS banner with copyright and version info
-// prettier-ignore
-const jsBanner = `${pckg.name} ${pckg.version} (${date})
-${pckg.homepage}
-@license ${pckg.license}`;
-const jsBannerPlugin = new webpack.BannerPlugin({
-    banner: jsBanner,
-    test: /\.js$/
-});
 
 // inject JS version number
 const jsVersionPlugin = new webpack.DefinePlugin({
@@ -56,5 +45,5 @@ module.exports = {
             }
         ]
     },
-    plugins: [jsBannerPlugin, jsVersionPlugin]
+    plugins: [jsVersionPlugin]
 };
